feat: Add post creation functionality
Adds functionality to create new posts on the platform, including sections and content of various types. This introduces a new `CreatePost` component and enhances the post model to support nested structures.
feat: Add endpoint to retrieve user's posts

Update build command

Update Package.json

Add SVG loader to project

This commit introduces the `vite-svg-loader` to the project, enabling the use of SVG images as components. This simplifies the process of adding SVGs to the application and offers advantages like:

- Improved performance through inlining SVG code.
- More flexibility for customizing SVGs with CSS classes.

The implementation also includes a fix for the opening of external links in a new tab, ensuring a better user experience.

Fix: Remove unnecessary legacy peer dependency flag

The `--legacy-peer-deps` flag was being used unnecessarily in the build script. This flag is no longer needed and can cause potential issues with newer versions of npm. Removing the flag simplifies the build process and ensures compatibility with current npm practices.